requirements.txt
flask
flask-cors
streamlit
requests
google-generativeai
